requirements.txt
alembic
bcrypt==3.2.2
cloudinary
django
fastapi
fastapi-limiter
fastapi-mail
faker
passlib[bcrypt]==1.7.4
pillow
psycopg2-binary
pyqrcode
python-dotenv
python-jose[cryptography]
python-multipart
pydantic
pydantic-settings
pytest
pytest-asyncio
qrcode
redis
sqlalchemy
sqlalchemy-utils
uvicorn
